requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dataframe-image==0.1.1
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mistune==0.8.4
mplfinance==0.12.7a17
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
numexpr==2.7.3
numpy==1.21.0
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
Pillow==8.3.0
pycodestyle==2.7.0
pycoingecko==2.2.0
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-telegram-bot==13.6
pytz==2021.1
# pywin32==301
pyzmq==22.1.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3
